requirements.txt
glfw==2.7.0
PyOpenGL==3.1.7
wgpu==0.15.1
PyGLM==2.7.1
trimesh==4.4.1
numpy>=1.26.0
scipy>=1.10.0